Remove implicit Sized bound from PhantomData's impls
Since PhantomData doesn't contain any data it doesn't matter if T is
Sized or not. There's an implicit Sized bound here that we have to
override to remove it.

This tweaks the (public) `impl_borrow_decode!(...)` and
`impl_borrow_decode_with_context!(...)` macros to support type bounds,
not just comma-separated token trees. I think this change should be
backwards-compatible.