Add support for #[bincode(skip)] to derive macros
Behaves like the serde equivalent. Upon decoding, this calls
`Default::default`. The attribute is valid on fields.

The serde version of `skip` is problematic, but only because of a bug in
serde, not because of a fundamental design problem:
- https://github.com/bincode-org/bincode/issues/184
- https://github.com/bincode-org/bincode/issues/424

This commit does not add skip for enum variants, only for struct and
enum fields.

Conditional skips (which this commit does not attempt to address) are
also problematic. We'd need to encode those values using a discriminant
or tombstone as if they were wrapped in Option:
- https://docs.rs/bincode/latest/bincode/spec/index.html#options