Add support for #[bincode(default = "...")] to derive macros
Behaves like the serde equivalent, though it's only ever valid in
combination with `skip`, as otherwise there's no reason for a field not
to be present in the serialized format.

This changes how `#[bincode(skip)]` works, so that it calls the provided
function instead of `Default::default()`.

This includes an error message if `default` is used without `skip`.